agent-sdk/requirements.txt
# Claude Agent SDK
anthropic>=0.40.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Date/time handling
pytz>=2024.1
python-dateutil>=2.8.0

# HTTP client
httpx>=0.27.0
requests>=2.31.0

# Database
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0

# Telegram notifications
python-telegram-bot>=20.0

# Alpaca trading
alpaca-py>=0.21.0

# Logging
structlog>=24.1.0
